+++ src/Service/TypeDossier/TypeDossierEditionService.php
@@ -3,6 +3,8 @@
 namespace Pastell\Service\TypeDossier;
 
 use Exception;
+use TypeDossierEtapeManager;
+use TypeDossierEtapeProperties;
 use TypeDossierException;
 use TypeDossierSQL;
 use TypeDossierPersonnaliseDirectoryManager;
@@ -27,10 +29,19 @@ class TypeDossierEditionService
     private $typeDossierPersonnaliseDirectoryManager;
 
     /**
+     * @var TypeDossierEtapeManager
+     */
+    private $typeDossierEtapeManager;
+
+    /**
      * @var TypeDossierExportService
      */
     private $typeDossierExportService;
 
+    /**
+     * @var TypeDossierManager
+     */
+    private $typeDossierManager;
     /**
      * @var Journal
      */
@@ -39,13 +50,30 @@ class TypeDossierEditionService
     public function __construct(
         TypeDossierSQL $typeDossierSQL,
         TypeDossierPersonnaliseDirectoryManager $typeDossierPersonnaliseDirectoryManager,
+        TypeDossierEtapeManager $typeDossierEtapeManager,
         TypeDossierExportService $typeDossierExportService,
-        Journal $journal
+        Journal $journal,
+        TypeDossierManager $typeDossierManager
     ) {
         $this->typeDossierSQL = $typeDossierSQL;
         $this->typeDossierPersonnaliseDirectoryManager = $typeDossierPersonnaliseDirectoryManager;
+        $this->typeDossierEtapeManager = $typeDossierEtapeManager;
         $this->typeDossierExportService = $typeDossierExportService;
         $this->journal = $journal;
+        $this->typeDossierManager = $typeDossierManager;
+    }
+
+    /**
+     * @param TypeDossierProperties $typeDossierProperties
+     * @return int
+     * @throws TypeDossierException
+     * @throws Exception
+     */
+    public function create(TypeDossierProperties $typeDossierProperties): int
+    {
+        $this->checkTypeDossierId($typeDossierProperties->id_type_dossier);
+        $this->checkNomOnglet($typeDossierProperties->nom_onglet);
+        return $this->edit(0, $typeDossierProperties);
     }
 
     /**
@@ -65,7 +93,6 @@ class TypeDossierEditionService
             $message_action = 'Modification';
         }
 
-        $this->checkTypeDossierId($typeDossierProperties->id_type_dossier);
         $typeDossierProperties = $this->fixSameStepsType($typeDossierProperties);
         $id_t = $this->typeDossierSQL->edit($id_t, $typeDossierProperties);
         $this->typeDossierPersonnaliseDirectoryManager->save($id_t, $typeDossierProperties);
@@ -79,6 +106,35 @@ class TypeDossierEditionService
             $message_action . " du type de dossier id_t=$id_t. JSON contenant l'export de la definition du type de dossier : " . $export
         );
         return $id_t;
+    }
+
+    /**
+     * @param string $source_type_dossier_id
+     * @param string $target_type_dossier_id
+     * @throws TypeDossierException
+     */
+    public function renameTypeDossierId(string $source_type_dossier_id, string $target_type_dossier_id)
+    {
+        $this->typeDossierPersonnaliseDirectoryManager->rename($source_type_dossier_id, $target_type_dossier_id);
+    }
+
+    /**
+     * @param $id_t
+     * @param $nom
+     * @param $type
+     * @param $description
+     * @param $nom_onglet
+     * @throws TypeDossierException
+     */
+    public function editLibelleInfo($id_t, $nom, $type, $description, $nom_onglet)
+    {
+        $this->checkNomOnglet($nom_onglet);
+        $typeDossierProporties = $this->typeDossierManager->getTypeDossierProperties($id_t);
+        $typeDossierProporties->nom = $nom;
+        $typeDossierProporties->type = $type;
+        $typeDossierProporties->description = $description;
+        $typeDossierProporties->nom_onglet = $nom_onglet;
+        $this->edit($id_t, $typeDossierProporties);
     }
 
     /**
@@ -109,6 +165,40 @@ class TypeDossierEditionService
             throw new TypeDossierException(
                 "L'identifiant du type de dossier « " . get_hecho($id_type_dossier) . " » ne doit pas dépasser " . self::TYPE_DOSSIER_ID_MAX_LENGTH . " caractères"
             );
+        }
+    }
+
+    /**
+     * @param $nom_onglet
+     * @throws TypeDossierException
+     */
+    public function checkNomOnglet($nom_onglet)
+    {
+        if (stristr($nom_onglet, '#')) {
+            throw new TypeDossierException(
+                "Le libellé de l'onglet principal ne doit pas contenir « # »"
+            );
+        }
+
+        $typeDossierEtape = new TypeDossierEtapeProperties();
+        $liste_onglet_name = [];
+        $all_etape_type = $this->typeDossierEtapeManager->getAllType();
+
+        foreach ($all_etape_type as $etape_type => $libelle) {
+            $typeDossierEtape->type = $etape_type;
+            foreach ($this->typeDossierEtapeManager->getFormulaireForEtape($typeDossierEtape) as $etape_onglet_name => $onglet_content) {
+                $liste_onglet_name[] = $etape_onglet_name;
+            }
+        }
+        $liste_onglet_name = array_unique($liste_onglet_name);
+        sort($liste_onglet_name);
+
+        foreach ($liste_onglet_name as $etape_onglet_name) {
+            if (strcasecmp($nom_onglet, $etape_onglet_name) == 0) {
+                throw new TypeDossierException(
+                    "Le libellé de l'onglet principal ne doit pas faire partie de la liste: " . implode(", ", $liste_onglet_name)
+                );
+            }
         }
     }