+++ tests/Service/TypeDossier/TypeDossierManagerTest.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace Pastell\Tests\Service\TypeDossier;
+
+use Pastell\Service\TypeDossier\TypeDossierManager;
+use Pastell\Service\TypeDossier\TypeDossierUtilService;
+use PastellTestCase;
+use TypeDossierProperties;
+
+class TypeDossierManagerTest extends PastellTestCase
+{
+    private function getTypeDossierManager()
+    {
+        return $this->getObjectInstancier()->getInstance(TypeDossierManager::class);
+    }
+
+    public function testGetEmptyDossierData()
+    {
+        $typeDossierData42 = $this->getTypeDossierManager()->getTypeDossierProperties(42);
+        $typeDossierData = new TypeDossierProperties();
+        $this->assertEquals($typeDossierData, $typeDossierData42);
+    }
+
+    public function testGetTypeDossierFromArray()
+    {
+        $json_content = json_decode(file_get_contents(TypeDossierImportServiceTest::FIXTURE_FILE), true);
+        $typeDossierProperties = $this->getTypeDossierManager()->getTypeDossierFromArray($json_content[TypeDossierUtilService::RAW_DATA]);
+        $this->assertEquals(TypeDossierImportServiceTest::ID_TYPE_DOSSIER, $typeDossierProperties->id_type_dossier);
+    }
+
+}